some restricted pages

+++ studymates/base/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from .models import *
 from .forms import RoomForm
 from django.db.models import Q
@@ -16,6 +17,8 @@ from django.contrib.auth.decorators import login_required
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -79,6 +82,9 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)  # returns a form that is prefilled
 
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -90,6 +96,9 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
     if request.method == "POST":
         room.delete()
         return redirect("home")
